refactor(html-artifacts): extract shared two-tier preview surface
Move StaticHtmlPreview, InteractiveHtmlPreview and their scroll/height
helpers from HtmlArtifactView.tsx (private) into
CodeBlockView/HtmlArtifactPreviewSurface.tsx, exported so the code-block
popup path can reuse them. The Outlet switches to the shared components
and the shared htmlArtifactPreviewRequiresInteractive gate (the same
expression it had inline). Behavior is unchanged — preparation for the
popup sandbox fix.
